adding right management for users

+++ cmail-api/users/user.php
@@ -34,6 +34,35 @@
 			return $this->endPoints;
 		}
 
+
+		private function getDelegated($write = true) {
+
+			$sql = "SELECT user_name, (user_authdata IS NOT NULL) AS user_login 
+				FROM users WHERE user_name IN (
+					SELECT api_delegate AS user_name FROM api_user_delegates WHERE api_user = :api_user
+				) OR user_name = :api_user";
+
+			$auth = Auth::getInstance($this->db, $this->output);
+
+			$params = array(
+				":api_user" => $auth->getUser()
+			);
+
+			$out = $this->db->query($sql, $params, DB::F_ARRAY);
+
+			$output = [];
+
+			foreach($out as $user) {
+				$user["modules"] = $this->getActiveModules($user);
+				$output[] = $user;
+			}
+
+			if ($write) {
+				$this->output->add("users", $output);
+			}
+			return $output;
+		}
+
 		/**
 		 * Returns the given user when in database. If no user is defined, send all users.
 		 * @param obj object with key username into
@@ -42,12 +71,42 @@
 		 */
 		public function get($obj, $write = true) {
 
-			if (!isset($obj["username"])) {
-				// if no username is set, return all users
-				return $this->getAll();
+			$auth = Auth::getInstance($this->db, $this->output);	
+
+			if ($auth->hasRight("admin")) {
+				if (!isset($obj["username"])) {
+					// if no username is set, return all users
+					return $this->getAll();
+				} else {
+					return $this->getByUser($obj, $write);
+				}
+			} else if ($auth->hasRight("delegate")) {
+				if (isset($obj["username"]) && !empty($obj["username"])) {
+					$users = $auth->getDelegateUsers();
+					$users[] = $auth->getUser();
+					foreach($users as $user) {
+						if ($user == $obj["username"]) {
+							return $this->getByUser($user);
+						}
+					}
+					$this->output->add("status", "User has no right to do this (not in delegated list).");
+					$this->output->add("users", []);
+					return [];
+				} else {
+					return $this->getDelegated($write);
+				}
+			} else {
+				if (!isset($obj["username"]) || $obj["username"] == $auth->getUser()) {
+					return $this->getByUser($auth->getUser());
+				} else {
+					$this->output->add("status","User has no right to do this (not the user).");
+					$this->output->add("users", []);
+					return [];
+				}
 			}
 
-			$username = $obj["username"];
+		}
+		private function getByUser($username, $write = true) {
 
 			$sql = "SELECT user_name, (user_authdata IS NOT NULL) AS user_login FROM users WHERE user_name = :user_name";
 		
@@ -116,6 +175,12 @@
 
 		}
 
+		/**
+		 * create a password hash.
+		 * @param $salt salt for the password. If null then a random one will taken.
+		 * @param $password the password
+		 * @return if salt is null then $salt:sha265($salt, $password), else sha265($salt, $password);
+		 */
 		public function create_password_hash($salt, $password) {
 
 
@@ -127,6 +192,24 @@
 			} else {
 				return hash("sha265", $salt . $password);
 			}
+		}
+
+		private function addDelegate($obj, $delegated = false) {
+			
+			$auth = Auth::getInstance($this->db, $this->output);
+
+			if (!$auth->hasRight("admin") && !$delegated) {
+				return false;
+			}
+
+			$sql = "INSERT INTO api_user_delegates (api_user, api_delegate) VALUES (:api_user, :api_delegate)";
+
+			$params = array(
+				":api_user" => $auth->getUser(),
+				":api_delegate" => $obj["api_delegate"]
+			);
+
+			return $this->db->insert($sql, [$params]);
 		}
 
 		/**
@@ -142,6 +225,12 @@
 
 			if (!isset($user["user_name"]) || empty($user["user_name"])) {
 				$this->output->add("status", "Username is not set.");
+				return false;
+			}
+
+			$auth = Auth::getInstance($this->db, $this->output);
+			if (!$auth->hasRight("admin") && !$auth->hasRight("delegate")) {
+				$this->output->add("status", "Not allowed");
 				return false;
 			}
 
@@ -165,6 +254,11 @@
 
 			if (isset($id) && !empty($id)) {
 				$this->output->add("user", true);
+
+				if ($auth->hasRight("delegate")) {
+					$this->addDelegate(array("api_delegate" => $user["user_name"]), true);
+				}
+
 				return true;
 			} else {
 				$this->output->add("user", false);
@@ -182,9 +276,24 @@
 		 */
 		public function set_password($user) {
 
-
+			$auth = Auth::getInstance($this->db, $this->output);
+			
+			$test = false;
 			if (!isset($user["user_name"]) || empty($user["user_name"])) {
 				$this->output->add("status", "Username is not set.");
+				return false;
+			}
+
+			if ($auth->hasRight("admin")) {
+				$test = true;
+			} else if ($auth->hasRight("delegate") && $auth->hasDelegatedUser($user["user_name"])) {
+				$test = true;
+			} else if ($auth->getUser() === $user["user_name"]) {
+				$test = true;	
+			}
+
+			if (!$test) {
+				$this->output->add("status", "Not allowed.");
 				return false;
 			}
 
@@ -219,12 +328,20 @@
 		 * @return false on error, else true
 		 */
 		public function delete($obj) {
-			//TODO: check input
-			
+			$auth = Auth::getInstance($this->db, $this->output);
 			if (!isset($obj["user_name"]) || empty($obj["user_name"])) {
 				$this->output->add("status", "No username set.");
 				return false;
-			}	
+			}
+			$test = false;
+			if ($auth->hasRight("admin") || ($auth->hasRight("delegate") && $auth->hasDelegatedUser($obj["user_name"]))) {
+				$test = true;
+			}
+
+			if (!$test && $auth->getUser() !== $obj["user_name"]) {
+				$this->output->add("status", "Not allowed.");
+				return false;
+			}
 			
 			$sql = "DELETE FROM users WHERE user_name = :username";
